# contracts/openapi.yaml
openapi: 3.1.0
info:
  title: HelpMate API
  version: 0.1.0
paths:
...
  /api/voice/transcribe:
    post:
      tags:
      - voice
      summary: Transcribe
      operationId: transcribe_api_voice_transcribe_post
      parameters:
      - name: language
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Language
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Transcript'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      requestBody:
        required: true
        content:
          audio/webm:
            schema:
              type: string
              format: binary
          audio/mp4:
            schema:
              type: string
              format: binary
          audio/ogg:
            schema:
              type: string
              format: binary
          audio/wav:
            schema:
              type: string
              format: binary
          audio/mpeg:
            schema:
              type: string
              format: binary
  /api/voice/speak:
    post:
      tags:
      - voice
      summary: Speak
      operationId: speak_api_voice_speak_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SpeakIn'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            audio/wav:
              schema:
                type: string
                format: binary
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
components:
  schemas:
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    SpeakIn:
      properties:
        text:
          type: string
          maxLength: 2000
          minLength: 1
          title: Text
        voice:
          anyOf:
          - type: string
          - type: 'null'
          title: Voice
      additionalProperties: false
      type: object
      required:
      - text
      title: SpeakIn
...
    Transcript:
      properties:
        text:
          type: string
          title: Text
        language:
          anyOf:
          - type: string
          - type: 'null'
          title: Language
        duration_ms:
          type: integer
          title: Duration Ms
        stt_ms:
          type: integer
          title: Stt Ms
      additionalProperties: false
      type: object
      required:
      - text
      - duration_ms
      - stt_ms
      title: Transcript
...
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
        input:
          title: Input
        ctx:
          type: object
          title: Context
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError
...